@@ -3,13 +3,12 @@ version: 0.0.0
 publish_to: none
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  build_runner: '>=1.10.1 <3.0.0'
-  build_test: '>=1.2.1 <3.0.0'
-  build_web_compilers: '>=2.9.0 <4.0.0'
-  meta: ">=1.2.2 <1.7.0" # Workaround to avoid https://github.com/dart-lang/sdk/issues/46142
-  test: ^1.15.7
+  build_runner: ^2.1.2
+  build_test: ^2.1.3
+  build_web_compilers: ^3.0.0
+  test: ^1.17.12
   test_html_builder:
     path: ../..
