@@ -7,38 +7,32 @@ description: >
 
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 executables:
   browser_aggregate_tests:
 
 dependencies:
-  args: '>=1.6.0 <3.0.0'
-  build: '>=1.3.0 <3.0.0'
-  dart_style: '>=1.3.6 <3.0.0'
-  glob: '>=1.2.0 <3.0.0'
-  json_annotation: ^3.1.1
-  path: ^1.7.0
-  test: ^1.15.7
-  test_core: ^0.3.11+4
-  yaml: ^2.2.1
+  args: ^2.3.0
+  build: ^2.1.0
+  dart_style: ^2.1.1
+  glob: ^2.0.1
+  json_annotation: ^4.1.0
+  path: ^1.8.0
+  test: ^1.17.12
+  test_core: ^0.4.2
+  yaml: ^3.1.0
 
 dev_dependencies:
-  build_runner: '>=1.10.1 <3.0.0'
-  build_test: '>=1.2.1 <3.0.0'
-  build_web_compilers: ^2.9.0
-  dependency_validator: ^2.0.1
-  meta: ">=1.2.2 <1.7.0" # Workaround to avoid https://github.com/dart-lang/sdk/issues/46142
-  pedantic: ^1.9.2
-  test_descriptor: ^1.2.0
-  test_process: ^1.0.5
+  build_runner: ^2.1.2
+  build_test: ^2.1.3
+  build_web_compilers: ^3.0.0
+  dependency_validator: ^3.1.0
+  pedantic: ^1.11.1
+  test_descriptor: ^2.0.0
+  test_process: ^2.0.2
 
   # If changes are made to `lib/src/config.dart` and regeneration is needed,
   # uncomment this and comment out the test_html_builder definition in
   # `build.yaml`.
-  # json_serializable: ^3.0.0
   # json_serializable: ^4.0.0
-
-dependency_validator:
-  ignore:
-    - meta
